Use uint8_t as BPF map value instead.

+++ src/IPIPBlock_loader.c
@@ -207,7 +207,7 @@ void UpdateList()
 
         // Convert IP to decimal and host byte order to store.
         uint32_t ip = inet_addr(line);
-        uint32_t val = 1;
+        uint8_t val = 1;
 
         // Attempt to update BPF map.
         if (bpf_map_update_elem(blacklist_map_fd, &ip, &val, BPF_ANY) != 0)
